dce: Split code
